+++ widget/src/providers/index.tsx
@@ -1,6 +1,11 @@
 import * as React from 'react'
 import HeroUiProvider from './HeroUiProvider'
+import ReownProvider from './ReownAppKit'
 
 export function Providers({ children }: { children: React.ReactNode }) {
-  return <HeroUiProvider>{children}</HeroUiProvider>
+  return (
+    <ReownProvider>
+      <HeroUiProvider>{children}</HeroUiProvider>
+    </ReownProvider>
+  )
 }
